+++ src/test/kotlin/hu/netsurf/erp/warehouse/testobject/CreateProductInputTestObject.kt
@@ -6,6 +6,7 @@ import hu.netsurf.erp.warehouse.constant.ProductTestConstants.PRODUCT_1_LONG_UNI
 import hu.netsurf.erp.warehouse.constant.ProductTestConstants.PRODUCT_1_NAME
 import hu.netsurf.erp.warehouse.constant.ProductTestConstants.PRODUCT_1_UNIT
 import hu.netsurf.erp.warehouse.input.CreateProductInput
+import java.math.BigDecimal
 
 class CreateProductInputTestObject {
     companion object {
@@ -26,7 +27,7 @@ class CreateProductInputTestObject {
         private fun build(
             name: String = PRODUCT_1_NAME,
             supplierId: Int = 1,
-            price: Double = 10000.0,
+            price: BigDecimal = BigDecimal(10000.00),
             unit: String = PRODUCT_1_UNIT,
             onStock: Int = 10,
         ): CreateProductInput =
